properly do attention in float

+++ nemo/collections/tts/modules/submodules.py
@@ -18,6 +18,8 @@ import torch
 from torch.autograd import Variable
 from torch.nn import functional as F
 
+# from nemo.utils import logging
+
 
 class LinearNorm(torch.nn.Module):
     def __init__(self, in_dim, out_dim, bias=True, w_init_gain='linear'):
@@ -135,19 +137,21 @@ class Attention(torch.nn.Module):
         attention_weights_cat: previous and cummulative attention weights
         mask: binary mask for padded data
         """
+        #     attention_hidden_state = attention_hidden_state.float()
+        #     processed_memory = processed_memory.float()
+        #     attention_weights_cat = attention_weights_cat.float()
+        #     memory = memory.float()
         with torch.cuda.amp.autocast(enabled=False):
-            attention_hidden_state = attention_hidden_state.float()
-            processed_memory = processed_memory.float()
-            attention_weights_cat = attention_weights_cat.float()
-            memory = memory.float()
-            alignment = self.get_alignment_energies(attention_hidden_state, processed_memory, attention_weights_cat)
+            alignment = self.get_alignment_energies(
+                attention_hidden_state.float(), processed_memory, attention_weights_cat.float()
+            )
 
             if mask is not None:
                 alignment.data.masked_fill_(mask, self.score_mask_value)
 
-            attention_weights = F.softmax(alignment, dim=1)
-            attention_context = torch.bmm(attention_weights.unsqueeze(1), memory)
-            attention_context = attention_context.squeeze(1)
+        attention_weights = F.softmax(alignment, dim=1)
+        attention_context = torch.bmm(attention_weights.unsqueeze(1), memory)
+        attention_context = attention_context.squeeze(1)
 
         return attention_context, attention_weights
